Map-Reduce/function-create-chunks/requirements.txt
functions-framework==3.*
google-cloud==0.34.0
google-cloud-storage==2.13.0
spacy==2.3.5
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz